pyproject.toml
[tool.poetry]
name = "fast-url"
version = "0.1.0"
description = ""
authors = ["Ali Shokrollahi <ali.shokrollahi.pro@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.30"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
aiosqlite = "^0.20.0"
base58 = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
